petland/extras/tools/art/requirements-tools.txt
Pillow==10.3.0
cairosvg==2.7.1
